requirements.txt
pycord==2.2.2
python-dotenv==0.20.0
aiohttp==3.8.3